Error handling improvements
